Output basics
Part 1
Write code that outputs the following. End with a newline (meaning use endl).
"This week was marvelous."

Part 2
Write code that outputs the following. End with a newline (meaning use endl).
"cot
pad"

Part 3
Write code that outputs the following. End with a newline (meaning use endl).
"Protect the
   birds"

Part 4
Write code that outputs variable numSeconds. End with a newline.

Part 5
If input is 3, output is:
Birds: 3

Part  6
Write code that outputs variable numCars as follows. End with a newline.
If input is 4, output is:
Were there 4 cars?